# Fix slow effects page query