Query in `EffectsQ.Page` method was performing a full-table scan to get results. Analyzing `EXPLAIN ANALYZE` output it seems that planner wasn't able to figure out which index to use to speed up the query. I rewrote the query to the equivalent query but that makes use of the index.

Close #314.

---
Old query:
```
horizon=> EXPLAIN ANALYZE SELECT heff.*, hacc.address FROM history_effects heff LEFT JOIN history_accounts hacc ON hacc.id = heff.history_account_id WHERE (
horizon(>            heff.history_operation_id > '69950027815981057'                                                           
horizon(>   OR (                                                                                              
horizon(>                           heff.history_operation_id = '69950027815981057'                                            
horizon(>           AND heff.order > '3'                                                                       
horizon(>   )) ORDER BY heff.history_operation_id asc, heff.order asc LIMIT 10;
                                                                             QUERY PLAN                                                                             
--------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.85..120.45 rows=10 width=181) (actual time=2674.930..2675.018 rows=10 loops=1)
   ->  Nested Loop Left Join  (cost=0.85..533414.63 rows=44600 width=181) (actual time=2674.929..2675.011 rows=10 loops=1)
         ->  Index Scan using hist_e_by_order on history_effects heff  (cost=0.43..492824.63 rows=44600 width=124) (actual time=2674.916..2674.924 rows=10 loops=1)
               Filter: ((history_operation_id > 69950027815981057::bigint) OR ((history_operation_id = 69950027815981057::bigint) AND ("order" > 3)))
               Rows Removed by Filter: 5999398
         ->  Index Scan using index_history_accounts_on_id on history_accounts hacc  (cost=0.42..0.90 rows=1 width=65) (actual time=0.006..0.007 rows=1 loops=10)
               Index Cond: (id = heff.history_account_id)
 Planning time: 0.399 ms
 Execution time: 2675.055 ms
(9 rows)
```

New query:
```
horizon=> EXPLAIN ANALYZE SELECT heff.*, hacc.address FROM history_effects heff LEFT JOIN history_accounts hacc ON hacc.id = heff.history_account_id WHERE (
horizon(>   (heff.history_operation_id >= '69950027815981057')
horizon(>   AND
horizon(>   (heff.history_operation_id > '69950027815981057' OR (heff.history_operation_id = '69950027815981057' AND heff.order > '3'))
horizon(> ) ORDER BY heff.history_operation_id asc, heff.order asc LIMIT 10;
                                                                            QUERY PLAN                                                                            
------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.85..2017.00 rows=10 width=181) (actual time=0.052..0.139 rows=10 loops=1)
   ->  Nested Loop Left Join  (cost=0.85..72380.38 rows=359 width=181) (actual time=0.051..0.132 rows=10 loops=1)
         ->  Index Scan using hist_e_by_order on history_effects heff  (cost=0.43..69634.83 rows=359 width=124) (actual time=0.027..0.036 rows=10 loops=1)
               Index Cond: (history_operation_id >= 69950027815981057::bigint)
               Filter: ((history_operation_id > 69950027815981057::bigint) OR ((history_operation_id = 69950027815981057::bigint) AND ("order" > 3)))
               Rows Removed by Filter: 3
         ->  Index Scan using index_history_accounts_on_id on history_accounts hacc  (cost=0.42..7.64 rows=1 width=65) (actual time=0.005..0.006 rows=1 loops=10)
               Index Cond: (id = heff.history_account_id)
 Planning time: 0.502 ms
 Execution time: 0.180 ms
(10 rows)

```